fix(FIN-021): fixed routes and logout bug
